scraper-service/requirements.txt
# Core scraping
scrapling>=0.2.0
playwright>=1.40.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
asyncpg>=0.29.0

# Scheduling
apscheduler>=3.10.4

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
